# lib/api-zod/src/generated/types/claimInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Campus Pick lost and found API
 * OpenAPI spec version: 0.1.0
 */

export interface ClaimInput {
  itemId: number;
  /** @minLength 2 */
  uniqueMarks: string;
  /** @nullable */
  contents?: string | null;
  /** @nullable */
  serialNumber?: string | null;
  /** @minLength 5 */
  additionalDetails: string;
}