# lib/api-zod/src/generated/types/handoverInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Campus Pick lost and found API
 * OpenAPI spec version: 0.1.0
 */

export interface HandoverInput {
  /** @minLength 4 */
  code: string;
}