# lib/api-zod/src/generated/types/item.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Campus Pick lost and found API
 * OpenAPI spec version: 0.1.0
 */
import type { ItemType } from './itemType';

export interface Item {
  id: number;
  reportId: string;
  type: ItemType;
  itemName: string;
  category: string;
  description: string;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  color?: string | null;
  location: string;
  eventDate: string;
  /** @nullable */
  approximateTime?: string | null;
  /** @nullable */
  condition?: string | null;
  status: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  identifyingFeatures?: string | null;
  /** @nullable */
  storageStatus?: string | null;
  submittedBy: string;
  createdAt: string;
  isSaved: boolean;
  /** @nullable */
  matchScore?: number | null;
  /** @nullable */
  matchLabel?: string | null;
}