# lib/api-zod/src/generated/types/itemDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Campus Pick lost and found API
 * OpenAPI spec version: 0.1.0
 */
import type { Item } from './item';
import type { Match } from './match';
import type { TimelineEvent } from './timelineEvent';

export type ItemDetail = Item & {
  matches?: Match[];
  timeline?: TimelineEvent[];
};