# lib/api-zod/src/generated/types/itemList.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Campus Pick lost and found API
 * OpenAPI spec version: 0.1.0
 */
import type { Item } from './item';

export interface ItemList {
  items: Item[];
  total: number;
}