# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Campus Pick lost and found API
 * OpenAPI spec version: 0.1.0
 */

export interface Message {
  id: number;
  threadId: string;
  senderName: string;
  senderRole: string;
  preview: string;
  body: string;
  createdAt: string;
  unread: boolean;
  /** @nullable */
  itemName?: string | null;
}