# lib/api-zod/src/generated/types/messageInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Campus Pick lost and found API
 * OpenAPI spec version: 0.1.0
 */

export interface MessageInput {
  threadId: string;
  /** @minLength 1 */
  body: string;
}