# lib/api-zod/src/generated/types/saveState.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Campus Pick lost and found API
 * OpenAPI spec version: 0.1.0
 */

export interface SaveState {
  itemId: number;
  saved: boolean;
}